fix(storage): resolve slot collision between governanceMultisig and s_l1FeeCoefficient
- Removed all unused `pending` variables.
- Previously, `governanceMultisig` in CommitReveal2 and `s_l1FeeCoefficient` in CommitReveal2L2
  overlapped in the same storage slot, causing all tests to fail.
- Changed `s_l1FeeCoefficient` type from `uint8` to `uint256` to align with EVM slot boundaries
  and eliminate storage collision.
- After this fix, all test cases now pass successfully.

+++ test/MultisigWallet.t.sol
@@ -191,7 +191,7 @@ contract MultisigWalletTest is Test {
     }
 
     function testL1FeeCoefficientGovernance() public {
-        uint8 newCoefficient = 75;
+        uint256 newCoefficient = 75;
 
         bytes memory callData =
             abi.encodeWithSelector(ICommitReveal2L2Governance.setL1FeeCoefficient.selector, newCoefficient);
